+++ sdk-lib/src/androidTest/java/cash/z/ecc/fixture/FakeRustBackend.kt
@@ -131,6 +131,10 @@ internal class FakeRustBackend(
         error("Intentionally not implemented yet.")
     }
 
+    override suspend fun pcztRequiresSaplingProofs(pczt: ByteArray): Boolean {
+        error("Intentionally not implemented yet.")
+    }
+
     override suspend fun addProofsToPczt(pczt: ByteArray): ByteArray {
         error("Intentionally not implemented yet.")
     }