+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/Synchronizer.kt
@@ -345,6 +345,17 @@ interface Synchronizer {
     suspend fun redactPcztForSigner(pczt: Pczt): Pczt
 
     /**
+     * Checks whether the caller needs to have downloaded the Sapling parameters.
+     *
+     * @param pczt The partially created transaction in its serialized format.
+     *
+     * @return `true` if this PCZT requires Sapling proofs.
+     *
+     * @throws PcztException.RedactPcztForSignerException as a common indicator of the operation failure
+     */
+    suspend fun pcztRequiresSaplingProofs(pczt: Pczt): Boolean
+
+    /**
      * Adds proofs to the given PCZT.
      *
      * @param pczt The partially created transaction in its serialized format.