+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/TypesafeBackendImpl.kt
@@ -156,6 +156,9 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
 
     override suspend fun redactPcztForSigner(pczt: Pczt): Pczt = Pczt(backend.redactPcztForSigner(pczt.toByteArray()))
 
+    override suspend fun pcztRequiresSaplingProofs(pczt: Pczt): Boolean =
+        backend.pcztRequiresSaplingProofs(pczt.toByteArray())
+
     override suspend fun addProofsToPczt(pczt: Pczt): Pczt = Pczt(backend.addProofsToPczt(pczt.toByteArray()))
 
     override suspend fun extractAndStoreTxFromPczt(