+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/transaction/TransactionEncoder.kt
@@ -101,6 +101,8 @@ internal interface TransactionEncoder {
 
     suspend fun redactPcztForSigner(pczt: Pczt): Pczt
 
+    suspend fun pcztRequiresSaplingProofs(pczt: Pczt): Boolean
+
     suspend fun addProofsToPczt(pczt: Pczt): Pczt
 
     suspend fun extractAndStoreTxFromPczt(